Add semi-working demo env where mui-tiptap updates can be reflected
This somewhat works, where updates to mui-tiptap's code will be able to
be reflected in the demo app without manually canceling the dev process,
rebuilding, and restarting the demo Vite server. However, it results in
a crash in the Tiptap editor:

```
Uncaught TypeError: Cannot read properties of null (reading 'updateOuterDeco')
    at EditorView.updateStateInner (index.js:5041:34)
    at EditorView.update (index.js:4973:14)
    at EditorView.setProps (index.js:4987:14)
    at Editor2.createNodeViews (Editor.ts:300:15)
    at PureEditorContent.init (EditorContent.tsx:66:14)
    at PureEditorContent.componentDidMount (EditorContent.tsx:41:10)
    at invokeLayoutEffectMountInDEV (react-dom.development.js:25133:22)
    at invokeEffectsInDev (react-dom.development.js:27351:11)
    at commitDoubleInvokeEffectsInDEV (react-dom.development.js:27327:5)
    at flushPassiveEffectsImpl (react-dom.development.js:27056:5)
```

Seems perhaps that's related to a bug in Tiptap more generally with
React.StrictMode compliance, since removing `StrictMode` allows the app
to refresh properly when mui-tiptap code changes and is rebuilt (and the
above stack trace specifically refers to double-invoking effects). We
could perhaps remove StrictMode, but it's nice to have it otherwise.