Improve build process to not bundle, and update example project script
We need to blow away the `example/node_modules` directory every time,
since it seems otherwise (whether we use `file:../` or `workspace:../`),
it will not use the latest version of the build and will instead
reference a cached version. Using `link:../` is always incorrect, as it
includes the entirety of the project and its dependencies, rather than
*installing* the package like a normal install, which leads to multiple
versions of react, separate tss-react MUI instances, etc.