Ensure ControlledBubbleMenu re-positions when clicking between links
Without this change, there was a minor bug here where if you click from
one link to another while using the LinkBubbleMenu, the original
getBoundingClientRect (and so original `from` and `to` values from this
outer scope) would be "cached" in the MUI Popper. As such the menu will
not move to the newly clicked element until clicking out and clicking
back again.