Remove sourceMap comments in TS definitions, since source not included
Otherwise, our built js files will reference source ts/tsx files that
are not actually included in the published npm package. (We currently
only publish the `dist/` dir.) We may want to publish the source in the
future, but for now, just don't include sourceMap references in the TS
files.