Update example app's dependencies
This adds back the bubble menu as a dependency, since otherwise because
it's a dependency of the tiptap/react plugin, it would get installed
with a higher version that had incompatibilities with our other
packages.

To get `react-hook-form` to be removed upon install, I had to delete the
pnpm-lock file for the example app in order to force it to re-install
the package based on its latest package.json definition. Otherwise, in
this case, the `react-hook-form` package was not getting removed, since
it seems the dependencies of the `file:../` for mui-tiptap were somehow
cached.