Don't cancel in-progress jobs for same branch
GitHub Actions allows 20 concurrent jobs as is
https://docs.github.com/en/actions/learn-github-actions/usage-limits-billing-and-administration#usage-limits,
which is probably more than we'll ever use. This avoids the overall
build being marked as "failing" (in the README CI badge) when pushing a
tag that runs a "redundant" version of the build-test job, and this
reduces noise from "build failed" emails that would be sent for any
canceled jobs.