Don't require `value` for MenuButton, defaulting to tooltipLabel
We don't actually *use* the `value` anywhere right now, it's just an
excessive field that's required by MUI for the purposes of controlling
MenuButtonGroups. So this makes it optional, with the default set as the
`tooltipLabel`, so it can still satisfy MUI (and could still be used in
a button group successfully), but still allows callers to use an
explicit `value` if they want.