Improve table bubble menu placement by affixing to top start
Don't have the bubble menu move around as much as the table is being
edited (columns and rows being added, removed, or resized). There are
some semi-hacky approaches taken here to ensure that we also do not
cover up a (potentially sticky) menu bar above the editor, and only fall
back to bottom positions if the top has no room.