Reduce default bubble menu elevation
Typically, bubble menus are interacted with alongside the content, so
with an `elevation={10}` in light mode in MUI, it would tend to cast a
relatively large shadow, which could be annoying (e.g. with the table
menu) if you're trying to edit the content that's directly next to the
menu. This adjusts it down slightly while still maintaining enough
elevation to be visually distinct from the editor content.