Lint against more extraneous dependencies being imported
No devDependencies should be imported in the main project without also
appearing in peerDependencies. Otherwise we risk the package breaking
when installed for others.

And because we publish our types with our package, even Type imports should
need to be included in dependencies or peerDependencies. So `type-fest`
should really be listed as a direct dependency (as it is for other major
users of type-fest, like yup, snakecase-keys, etc., and as described
here
https://github.com/sindresorhus/type-fest/issues/136#issuecomment-714211609).