Loosen peer deps for tiptap packages
Let consumers of the app install newer versions of these various
packages if they so choose. It would be nice if this could include
packages with versions above 2.0.0-beta.210, but due to
https://tiptap.dev/blog/new-pm-package-and-upgrade-guide-for-beta-210
and in particular a change in the table extension, which went from using
`@tiptap/prosemirror-tables` to the original `prosemirror-tables`
package (via `@tiptap/pm`), so we can't easily support both.

We'll bump all tiptap packages to beta.210+ and switch to tiptap/pm
separately later.