Fix link bubble menu state when clicking out or changing selection
This avoids a focus-jump when clicking away from the "Edit" form, by
ensuring the bubble menu state is kept in sync. An extra precautionary
fallback to `?? ""` is added for the `currentHref` just in case the menu
renders when unexpected and no `href` is available, since this can cause
a crash (though that scenario is likely resolved by our more deliberate
menu-closing, and handled here just as a defensive precaution).