Use editor element directly as container, and allow overrides
We no longer clip when just using the `editor.element` directly, even
with the OutlinedField as a parent, so this approach is less hacky and
cleaner (and avoids some quirks where insertion at a parent node could
cause problems).

And now users can override it (including reverting back to the default
MUI behavior by passing `null`) if they so choose.