Don't set a default bubble menu container, and allow disablePortal
It turns out that when closing a MUI Dialog that contains a
ControlledBubbleMenu where `container={editor.options.element}`, it will
lead to:

```
Uncaught DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
```

This is probably likely due to some quirk of how Tiptap's elements are
rendered. Setting to `editor.options.element.parentElement` instead
tends to avoid the error, but in either case, it's actually not an ideal
DOM placement. This is because with a MUI `<Dialog />` for instance, the
Popper will still be inside the inner `Paper` and `MuiDialog-container`,
where it will restrict the positioning of the bubble menu to stay
contained in the same dimensions as the overall Dialog, rather than
allowing the bubble menu to go beyond it. When the editor and dialog are
small, this can lead to scroll bars being added and a janky UI, and even
when the editor is larger, it can end up overlapping editor content in
ways that wouldn't happen outside a modal. By contrast, setting to
`container` prop to the `<Dialog />`'s `ref` value will avoid the errors
mentioned above and also allow the bubble menu to appear as it more
naturally would, while still being "on top" of the modal.