Ignore Tiptap links in markdown-link-check in README
Tiptap's site is timing out a lot of the time, leading to spurious build
failures like
https://github.com/sjdemartini/mui-tiptap/actions/runs/5558982096/jobs/10154638405#step:11:80

It's also slowing down the CI runs. So better to just ignore them.

The main reason we have markdown-link-check is for local markdown links
(not external web pages), so this was only a nice to have.