Remove @tiptap/extension-text-style as peer dep
We only need to reference the types from that extension locally (not
exported, and not used directly), so we can remove a peer dependency on
it and a direct import in our `FontSize` extension. Users of the
`FontSize` extension were already required to separately install the
`TextStyle` extension and include it in their `extensions` array, so
they will not encounter problems importing and using `FontSize` when
it's not listed as a peer dep.