Remove Tiptap Link extension from peer deps
We don't actually directly depend on the code being imported anywhere
within mui-tiptap, so it need not be listed as an explicit peer
dependency. Users should still install `@tiptap/extension-link` if they
intend to use the `Link` extension themselves and the associated
mui-tiptap components and extensions (`LinkBubbleMenu`,
`LinkBubbleMenuHandler`, etc.) as those depend on the `Link` extension
being present.