Show separate color indicator bar in text & highlight buttons
This adds generic support for showing the color indicator in the base
`MenuButtonColorPicker` component, and utilizes that in
`MenuButtonTextColor` and `MenuButtonHighlightColor`.

Now, the current color will be shown with the "bar" that was present in
the original icons. The bar will be omitted (effectively transparent) if
no color/highlight value is active.