Fall back to default mark color if highlight is active but has "" color
If highlighting is active but the Tiptap state has color is empty, we
should fall back to showing the "active color" in button and
color-picker as the defaultMarkColor. We were already doing this if the
color was `null`, but this fixes it so it behaves that way if the
`color` attribute is an empty string too.